% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{doi2handle}
\alias{doi2handle}
\alias{handle2doi}
\title{Get a Dryad DOI from a handle, and vice versa}
\usage{
doi2handle(...)

handle2doi(...)
}
\arguments{
\item{...}{ignored}
}
\description{
These functions are defunct
}
\keyword{internal}
