% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doi2handle.R
\name{doi2handle}
\alias{doi2handle}
\alias{handle2doi}
\title{Get a Dryad DOI from a handle, and vice versa}
\usage{
doi2handle(x, ...)

handle2doi(x, ...)
}
\arguments{
\item{x}{(character) A Dryad dataset DOI or handle. required}

\item{...}{Curl options, passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
(character) a DOI or handle
}
\description{
Get a Dryad DOI from a handle, and vice versa
}
\examples{
\dontrun{
doi2handle('10.5061/dryad.c0765')
handle2doi('10255/dryad.153920')

doi2handle('10.5061/dryad.c0765')
}
}
