% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_url.r
\name{download_url}
\alias{download_url}
\title{Get a URL given a Dryad ID}
\usage{
download_url(doi = NULL, handle = NULL, id = NULL, ...)
}
\arguments{
\item{doi}{(character) A Dryad dataset DOI, of the form
10.5061/dryad.xxx}

\item{handle}{(character) A Dryad dataset handle, of the form
10255/dryad.xxx}

\item{id}{Dryad identifier, i.e. '10255/dryad.19'. DEFUNCT, USE
\code{doi} or \code{handle} parameters}

\item{...}{Curl options, passed on to \code{\link[httr]{GET}}}
}
\value{
A URL for dataset for the Dryad id.
}
\description{
Get a URL given a Dryad ID
}
\examples{
\dontrun{
# look up by doi
download_url(doi = '10.5061/dryad.k7455')
# download_url(doi = '10.5061/dryad.9t0n8/1') # doesn't work

# look up by legacy handle
download_url(handle = '10255/dryad.1759')
download_url(handle = '10255/dryad.102551')
}
}
