% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_get_records.R
\name{dr_get_records}
\alias{dr_get_records}
\title{Download metadata for individual Dryad id's.}
\usage{
dr_get_records(ids, prefix = "oai_dc", as = "df", ...)
}
\arguments{
\item{id}{Dryad identifier, i.e. oai:datadryad.org:10255/dryad.8820}

\item{url}{the base url for the function (should be left to default).}
}
\value{
Metadata for a Dryad dataset.
}
\description{
Download metadata for individual Dryad id's.
}
\examples{
\dontrun{
dr_get_records(ids = 'oai:datadryad.org:10255/dryad.8820')
}
}

