% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_listidentifiers.r
\name{dr_list_identifiers}
\alias{dr_list_identifiers}
\title{Gets all OAI Dryad identifiers.}
\usage{
dr_list_identifiers(prefix = "oai_dc", from = NULL, until = NULL,
  set = "hdl_10255_3", token = NULL, as = "df", ...)
}
\arguments{
\item{...}{optional additional curl options (debugging tools mostly)}

\item{tor}{Return list of identifiers to R ('r'), or to your directory
at '~/.' ('dir') (character).}

\item{url}{the base url for the function (should be left to default).}

\item{curl}{If using in a loop, call getCurlHandle() first and pass
the returned value in here (avoids unnecessary footprint)}
}
\value{
List of OAI identifiers for each dataset.
}
\description{
Gets all OAI Dryad identifiers.
}
\examples{
\dontrun{
dr_list_identifiers(from='2010-01-01', until = "2010-06-30")
dr_list_identifiers(set="geocode1", from='2015-09-01', until='2015-09-05')
dr_list_identifiers(prefix="iso19139", from='2015-09-01', until='2015-09-20')
dr_list_identifiers(prefix="dif", from='2015-09-01', until='2015-09-20')

identifiers <- dr_list_identifiers('r')

# Data packages
identifiers[[1]]

# Data files
identifiers[[2]]
}
}

