% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_listmetadataformats.r
\name{dr_list_metadata_formats}
\alias{dr_list_metadata_formats}
\title{Learn about the Dryad metadata formats available.}
\usage{
dr_list_metadata_formats(...)
}
\value{
List of information on metadata formats.
}
\description{
Learn about the Dryad metadata formats available.
}
\examples{
\dontrun{
dr_list_metadata_formats()
}
}

