% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_listsets.r
\name{dr_list_sets}
\alias{dr_list_sets}
\title{List the sets in the Dryad metadata repository.}
\usage{
dr_list_sets(token = NULL, as = "df", ...)
}
\description{
Retrieve the set structure of Dryad, useful for selective harvesting
}
\examples{
\dontrun{
dr_list_sets()
dr_list_sets(as = "list")
dr_list_sets(as = "raw")
library('httr')
res <- dr_list_sets(config = verbose())
}
}

