% Generated by roxygen2 (4.0.2): do not edit by hand
\name{dr_listmetadataformats}
\alias{dr_listmetadataformats}
\title{Learn about the Dryad metadata formats available.}
\usage{
dr_listmetadataformats()
}
\value{
List of information on metadata formats.
}
\description{
Learn about the Dryad metadata formats available.
}
\examples{
\dontrun{
dr_listmetadataformats()
}
}

