% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_listsets.r
\name{dr_listsets}
\alias{dr_listsets}
\title{List the sets in the Dryad metadata repository.}
\usage{
dr_listsets()
}
\description{
Retrieve the set structure of Dryad, useful for selective harvesting
}
\examples{
\dontrun{
dr_listsets()
}
}

