% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dryad_add_file}
\alias{dryad_add_file}
\title{Upload a file for an in-progress Dryad dataset}
\usage{
dryad_add_file(doi, file, ...)
}
\arguments{
\item{doi}{a DOI for a dataset, of the form \code{10.1234/dryad.abcd},
without \verb{doi:}. required}

\item{file}{a file path; must exist. required}

\item{...}{curl options}
}
\value{
a list of metadata about the created file
}
\description{
Upload a file for an in-progress Dryad dataset
}
\details{
See \code{\link[=dryad_create_dataset]{dryad_create_dataset()}} for examples
}
