% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dryad_auth_test}
\alias{dryad_auth_test}
\title{Check that a Dryad token works}
\usage{
dryad_auth_test(token)
}
\arguments{
\item{token}{(character) a Dryad OAUth token. required}
}
\value{
errors if token is expired/invalid; returns named list on
success with slots: \code{message}, \code{user_id}
}
\description{
Check that a Dryad token works
}
\examples{
# dryad_auth_test("some-token")
}
