% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dryad_create_dataset}
\alias{dryad_create_dataset}
\alias{dryad_update_dataset}
\title{Create a Dryad dataset}
\usage{
dryad_create_dataset(meta, ...)

dryad_update_dataset(doi, meta, ...)
}
\arguments{
\item{meta}{an R6 object of class \code{DryadMetadata}}

\item{...}{curl options}

\item{doi}{a DOI for a dataset, of the form \code{10.1234/dryad.abcd},
without \verb{doi:}. required}
}
\value{
a list of metadata about the created dataset
}
\description{
Create a Dryad dataset
}
\examples{
\dontrun{
library(charlatan)
lorem <- charlatan::LoremProvider$new()
title <- lorem$sentence(nb_words = 10)
abstract <- lorem$paragraph(nb_sentences = 10)
name <- strsplit(ch_name(), "\\\\s")[[1]]
authors <- list(
  list(
    firstName = name[1],
    lastName = name[2],
    email = "myrmecocystus@gmail.com",
    affiliation = "The University"
  )
)

# create metadata
meta <- dryad_create_metadata(
  title = title, abstract = abstract, authors = authors)

# set base url to staging server
Sys.setenv(DRYAD_BASE_URL="https://dryad-stg.cdlib.org")

# create dataset
created_dataset <- dryad_create_dataset(meta)
created_dataset
doi <- created_dataset$doi

# create file on the dataset
file <- file.path(tempdir(), "some-data.csv")
write.table(iris, file = file, sep = ",", row.names = FALSE)
added_file <- dryad_add_file(doi, file)
added_file

# create file by url reference on the dataset
# FIXME: have not got this to work yet
url = "https://raw.githubusercontent.com/sckott/habanero/master/setup.py"
added_file_by_url <- dryad_add_by_url(doi = created_dataset$doi, url,
  description = "A file with some things and stuff", mimeType = "text/plain",
  path = "setup.py")

# view the dataset
# toggle env var to make sure your token is used
Sys.setenv(DRYAD_USE_TOKEN=TRUE)
dryad_dataset(doi)

# mark the dataset as submitted
dryad_submit(doi)

# update the dataset
# FIXME: haven't got this to work yet, just
# got: "Your dataset cannot be updated now" 
new_title <- lorem$sentence(nb_words = 10)
meta <- dryad_create_metadata(
  title = new_title, abstract = abstract, authors = authors)
dryad_dataset(doi)
dryad_update_dataset(doi, meta)

# delete a file - FIXME: not working yet
# dryad_delete_file()
}
}
