% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryad_metadata.R
\name{dryad_create_metadata}
\alias{dryad_create_metadata}
\title{Create Dryad metadata for a dataset}
\usage{
dryad_create_metadata(...)
}
\arguments{
\item{...}{metadata fields, must be named variables. For available fields
see Details.}
}
\description{
Create Dryad metadata for a dataset
}
\details{
The available metadata fields:
\itemize{
\item title: string (required)
\item abstract: string (required)
\item authors: list of named lists (required). required for each author:
\code{firstName}, \code{lastName}, \code{email}, \code{affiliation}
\item relatedWorks: list of named lists
\item funders: list of named lists
\item methods: string
\item usageNotes: string
\item keywords: vector of strings
\item locations: xxx
\item publicationISSN: string
\item publicationName: string
\item manuscriptNumber: string
\item identifier: string
\item userId: string
\item invoiceId: string
\item skipDataciteUpdate: boolean
\item skipEmails: boolean
\item loosenValidation: boolean
\item preserveCurationStatus: boolean
}
}
\examples{
# bad
# dryad_create_metadata(title = "Foo bar",
#   abstract = "And then and stuff", authors = "Adsf")
# dryad_create_metadata(title = TRUE,
#   abstract = "And then and stuff", authors = "Adsf")
# dryad_create_metadata(title = "Foo bar", abstract = 5,
#   authors = "Adsf")

# good
x <- dryad_create_metadata(
  title = "Foo bar", abstract = "And then and stuff",
  authors = list(
    list(
      firstName = "Wanda", 
      lastName = "Jackson",
      email = "wanda.jackson@example.com", 
      affiliation = "University of the Example"
    )
  )
)
x
x$to_json()
}
\references{
https://github.com/CDL-Dryad/dryad-app/blob/main/documentation/apis/dataset_metadata.md
}
