% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dryad_dataset}
\alias{dryad_dataset}
\title{Get datasets by DOI(s)}
\usage{
dryad_dataset(dois, ...)
}
\arguments{
\item{dois}{(character) one or more DOIs, required}

\item{...}{Further args passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of lists, each named by the input DOI
}
\description{
Get datasets by DOI(s)
}
\examples{
\dontrun{
dryad_dataset(doi = "10.5061/dryad.f385721n")
dois <- c("10.5061/dryad.f385721n", "10.5061/dryad.7ct1n", "10.5061/dryad.1g626")
dryad_dataset(dois = dois)
}
}
\seealso{
Other dryad-datasets: 
\code{\link{dryad_dataset_versions}()},
\code{\link{dryad_datasets}()}
}
\concept{dryad-datasets}
