% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dryad_dataset_versions}
\alias{dryad_dataset_versions}
\title{Get dataset versions by DOI(s)}
\usage{
dryad_dataset_versions(dois, ...)
}
\arguments{
\item{dois}{(character) one or more DOIs, required}

\item{...}{Further args passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of lists, each named by the input DOI
}
\description{
Get dataset versions by DOI(s)
}
\examples{
\dontrun{
x = dryad_dataset_versions(dois = "10.5061/dryad.f385721n")
x
dois <- c("10.5061/dryad.f385721n", "10.5061/dryad.7ct1n", "10.5061/dryad.1g626")
dryad_dataset_versions(dois = dois)
}
}
\seealso{
Other dryad-datasets: 
\code{\link{dryad_datasets}()},
\code{\link{dryad_dataset}()}
}
\concept{dryad-datasets}
