% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dryad_datasets}
\alias{dryad_datasets}
\title{List datasets}
\usage{
dryad_datasets(...)
}
\arguments{
\item{...}{Named key-value sequeunce of parameters passed to Dryad query
such as "page" (nunber), or "per_page".}
}
\value{
a tibble
}
\description{
List datasets
}
\examples{
\dontrun{
(x <- dryad_datasets())
x$meta
x$links
x$data
x <- dryad_datasets(per_page = 50, page = 7)
}
}
\seealso{
Other dryad-datasets: 
\code{\link{dryad_dataset_versions}()},
\code{\link{dryad_dataset}()}
}
\concept{dryad-datasets}
