% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dryad_delete_file}
\alias{dryad_delete_file}
\title{Delete a file for an in-progress Dryad dataset}
\usage{
dryad_delete_file(id, ...)
}
\arguments{
\item{id}{a file id. required}

\item{...}{curl options}
}
\value{
a list of metadata
}
\description{
Delete a file for an in-progress Dryad dataset
}
\details{
See \code{\link[=dryad_create_dataset]{dryad_create_dataset()}} for examples
}
