% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryad_download.R
\name{dryad_download}
\alias{dryad_download}
\title{Get datasets download by DOI(s)}
\usage{
dryad_download(doi, ...)
}
\arguments{
\item{...}{Further args passed on to \link[crul:verb-GET]{crul::verb-GET}}

\item{dois}{(character) one DOI, required}
}
\value{
a list of lists, each named by the input DOI
}
\description{
Get datasets download by DOI(s)
}
\examples{
\dontrun{
dryad_download(doi = "10.5061/dryad.f385721n")
z = dryad_download(doi = "10.5061/dryad.7ct1n", verbose = TRUE)
dois <- c("10.5061/dryad.f385721n", "10.5061/dryad.7ct1n", "10.5061/dryad.1g626")
dryad_download(dois = dois)
}
}
