% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryad_fetch.R
\name{dryad_fetch}
\alias{dryad_fetch}
\title{Download Dryad files}
\usage{
dryad_fetch(url, destfile = NULL, ...)
}
\arguments{
\item{url}{Dryad URL for a dataset.}

\item{destfile}{Destination file. If not given, we assign a file name based
on URL provided.}

\item{...}{Further args passed on to \code{\link{download.file}}}
}
\value{
A path to the file
}
\description{
Download Dryad files
}
\details{
This function is a thin wrapper around download.file to get files
to your machine only. We don't attempt to read/parse them in to R.
}
\examples{
\dontrun{
url <- download_url('10255/dryad.1759')
dryad_fetch(url)
}
}

