% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_url.r
\name{dryad_files}
\alias{dryad_files}
\title{Get a URL given a Dryad ID}
\usage{
dryad_files(doi, ...)
}
\arguments{
\item{doi}{(character) A Dryad dataset DOI, of the form
10.5061/dryad.xxx. required}

\item{...}{Curl options, passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
(character) One or more URLS for direct download of datasets
for the given Dryad DOI
}
\description{
Get a URL given a Dryad ID
}
\examples{
\dontrun{
# look up by doi
dryad_files(doi = '10.5061/dryad.1758')
dryad_files(doi = '10.5061/dryad.60699')
}
}
