% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{dryad_get_cache}
\alias{dryad_get_cache}
\title{Get 'rdryad' cache data as a \pkg{hoardr} object}
\usage{
dryad_get_cache()
}
\value{
A \pkg{hoardr} object defining the cache directory.
}
\description{
Used primarily to modify the cache directory and set with
\link{dryad_set_cache}.
}
