% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{dryad_package_dois}
\alias{dryad_package_dois}
\title{Get file DOIs for a Dryad package DOI}
\usage{
dryad_package_dois(doi, ...)
}
\arguments{
\item{...}{ignored}
}
\description{
This function is defunct
}
\keyword{internal}
