% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryad_package_dois.R
\name{dryad_package_dois}
\alias{dryad_package_dois}
\title{Get file DOIs for a Dryad package DOI}
\usage{
dryad_package_dois(doi, ...)
}
\arguments{
\item{doi}{(character) A Dryad package DOI. required}

\item{...}{Further args passed on to \link[crul:HttpClient]{crul::HttpClient}}
}
\value{
(character) zero or more DOIs for the files; if
no results a zero length character vector
}
\description{
Get file DOIs for a Dryad package DOI
}
\examples{
\dontrun{
dryad_package_dois('10.5061/dryad.1758')
dryad_package_dois('10.5061/dryad.9t0n8')
dryad_package_dois('10.5061/dryad.60699')
}
}
