% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{dryad_set_cache}
\alias{dryad_set_cache}
\title{Set the 'rdryad' cache by passing a \pkg{hoardr} object defining cache data
including local path.}
\usage{
dryad_set_cache(obj)
}
\arguments{
\item{obj}{A \pkg{hoardr} object returned from \link{dryad_get_cache}.}
}
\value{
(Invisibly) The update cache object.
}
\description{
Set the 'rdryad' cache by passing a \pkg{hoardr} object defining cache data
including local path.
}
