% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dryad_submit}
\alias{dryad_submit}
\title{Mark a Dryad dataset as submitted}
\usage{
dryad_submit(doi, ...)
}
\arguments{
\item{doi}{a DOI for a dataset, of the form \code{10.1234/dryad.abcd},
without \verb{doi:}. required}

\item{...}{curl options}
}
\value{
a list of metadata about the created file
}
\description{
Mark a Dryad dataset as submitted
}
\details{
other optional fields to include:
\itemize{
\item digest/digestType: (character) if they are added then they will be
passed as part of the ingest manifest to Merritt. If the digest doesn't
match when Merritt downloads the files from the internet, then Merritt
will cause an error on ingesting and you'll need to check/fix it
\item description: (character) brief description of the file
\item size: (numeric/integer) file size in bytes
\item path: (character) can provide a filename when the name is not specified
in the URL (this is common when the URL is using an identifier string
rather than a file name)
\item mimeType: (character) the mime type
\item skipValidation: (boolean) if \code{TRUE}, will tell DASH to skip the step
of validating the existence of the file
}

See \code{\link[=dryad_create_dataset]{dryad_create_dataset()}} for examples
}
