% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth.R
\name{dryad_token}
\alias{dryad_token}
\title{Get a Dryad Token via OAuth client id and secret}
\usage{
dryad_token(id, secret, ...)
}
\arguments{
\item{id}{(character) Dryad client id. required}

\item{secret}{(character) Dryad client secret. required}

\item{...}{curl options passed on to \link[crul:verb-POST]{crul::verb-POST}}
}
\value{
a named list: access_token, token_type, expires_in, scope,
and created_at
}
\description{
This does not do the OAUth dance but rather fetches a
token using your id and secret. Note that the token expires
after 10 hours. See link below for how to get a client id
and secret.
}
\details{
For more details on authentication see
https://github.com/CDL-Dryad/dryad-app/blob/main/documentation/apis/submission.md
}
\examples{
\dontrun{
id <- Sys.getenv('DRYAD_CLIENT_ID')
secret <- Sys.getenv('DRYAD_CLIENT_SECRET')
out <- dryad_token(id, secret)
out
# test that your auth token works
dryad_auth_test(out$access_token)
}
}
