% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryad_versions.R
\name{dryad_versions}
\alias{dryad_versions}
\alias{dryad_versions_files}
\title{Get a dataset version by version ID}
\usage{
dryad_versions(ids, ...)

dryad_versions_files(ids, ...)
}
\arguments{
\item{ids}{(character) one or version ids, required}

\item{...}{Further args passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of lists, each named by the input DOI
}
\description{
Get a dataset version by version ID
}
\examples{
\dontrun{
dryad_versions(ids = 18774)
dryad_versions_files(ids = 18774)
}
}
