\name{identify}
\alias{identify}
\title{Learn about the Dryad OAI-PMH service.}
\usage{
  identify(formatted = FALSE)
}
\arguments{
  \item{formatted}{Defaults to printing results - set to
  TRUE to silence printing.}
}
\value{
  List of information describing Dryad.
}
\description{
  Learn about the Dryad OAI-PMH service.
}
\examples{
\dontrun{
identify()
}
}

