\name{listmetadataformats}
\alias{listmetadataformats}
\title{Learn about the Dryad metadata formats available.}
\usage{
  listmetadataformats()
}
\value{
  List of information on metadata formats.
}
\description{
  Learn about the Dryad metadata formats available.
}
\examples{
\dontrun{
listmetadataformats()
}
}

