% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{oai-defunct}
\alias{oai-defunct}
\alias{dr_get_records}
\alias{dr_identify}
\alias{dr_list_records}
\alias{dr_list_identifiers}
\alias{dr_list_metadata_formats}
\alias{dr_list_sets}
\title{Defunct OAI-PMH functions}
\usage{
dr_get_records(...)

dr_identify(...)

dr_list_records(
  prefix = "oai_dc",
  from = NULL,
  until = NULL,
  set = "hdl_10255_3",
  token = NULL,
  as = "df",
  ...
)

dr_list_identifiers(
  prefix = "oai_dc",
  from = NULL,
  until = NULL,
  set = "hdl_10255_3",
  token = NULL,
  as = "df",
  ...
)

dr_list_metadata_formats(...)

dr_list_sets(token = NULL, as = "df", ...)
}
\arguments{
\item{...}{ignored}
}
\description{
Defunct OAI-PMH functions

Download metadata for individual Dryad id's

Learn about the Dryad OAI-PMH service.

List Dryad records

Gets OAI Dryad identifiers

Get available Dryad metadata formats

List the sets in the Dryad metadata repository.
}
\details{
The Dryad OAI-PMH service is no longer being updated
See http://wiki.datadryad.org/Old:Dryad_API#OAI-PMH

Defunct functions:
\itemize{
\item \code{dr_get_records}
\item \code{dr_identify}
\item \code{dr_list_records}
\item \code{dr_list_identifiers}
\item \code{dr_list_metadata_formats}
\item \code{dr_list_sets}
}
}
\keyword{internal}
