% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdryad-package.R
\docType{package}
\name{rdryad-package}
\alias{rdryad}
\alias{rdryad-package}
\title{Interface to the Solr API for Dryad, and the API to get datasets.}
\description{
Interface to the Solr API for Dryad, and the API to get datasets.
}

