% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryad_solr.r
\name{solr-defunct}
\alias{solr-defunct}
\alias{d_solr_search}
\alias{d_solr_facet}
\alias{d_solr_group}
\alias{d_solr_highlight}
\alias{d_solr_mlt}
\alias{d_solr_stats}
\title{Defunct Solr functions}
\usage{
d_solr_search(..., proxy = NULL, callopts = list())

d_solr_facet(..., proxy = NULL, callopts = list())

d_solr_group(..., proxy = NULL, callopts = list())

d_solr_highlight(..., proxy = NULL, callopts = list())

d_solr_mlt(..., proxy = NULL, callopts = list())

d_solr_stats(..., proxy = NULL, callopts = list())
}
\arguments{
\item{...}{ignored}
}
\description{
Defunct Solr functions

Search the Dryad Solr endpoint
}
\details{
The Dryad Solr service is no longer being updated
See http://wiki.datadryad.org/Old:Dryad_API#SOLR_search_access

Defunct functions:
\itemize{
\item \code{d_solr_search}
\item \code{d_solr_facet}
\item \code{d_solr_group}
\item \code{d_solr_highlight}
\item \code{d_solr_mlt}
\item \code{d_solr_stats}
}
}
\keyword{internal}
