% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{download_dryadmetadata}
\alias{download_dryadmetadata}
\title{Download metadata for individual Dryad id's}
\usage{
download_dryadmetadata(...)
}
\description{
This function changed name to \code{\link{dr_get_records}}
}
\keyword{internal}

