% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_url.r
\name{download_url}
\alias{download_url}
\title{Get a URL given a Dryad ID}
\usage{
download_url(id, ...)
}
\arguments{
\item{id}{Dryad identifier, i.e. '10255/dryad.19'.}

\item{...}{Curl options, passed on to \code{\link[httr]{GET}}}
}
\value{
A URL for dataset for the Dryad id.
}
\description{
Get a URL given a Dryad ID
}
\examples{
\dontrun{
download_url(id = '10255/dryad.1759')
download_url(id = '10255/dryad.102551')
}
}

