% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_identify.r
\name{dr_identify}
\alias{dr_identify}
\title{Learn about the Dryad OAI-PMH service.}
\usage{
dr_identify(...)
}
\arguments{
\item{...}{Curl debugging options passed on to \code{httr::GET}}
}
\value{
List of information describing Dryad.
}
\description{
Learn about the Dryad OAI-PMH service.
}
\examples{
\dontrun{
dr_identify()
}
}
