% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_listidentifiers.r
\name{dr_list_identifiers}
\alias{dr_list_identifiers}
\title{Gets OAI Dryad identifiers}
\usage{
dr_list_identifiers(prefix = "oai_dc", from = NULL, until = NULL,
  set = "hdl_10255_3", token = NULL, as = "df", ...)
}
\value{
XML character string, data.frame, or list, depending on what requested
witht the \code{as} parameter

List of OAI identifiers for each dataset.
}
\description{
Gets OAI Dryad identifiers
}
\examples{
\dontrun{
dr_list_identifiers(from='2010-01-01', until = "2010-06-30")
dr_list_identifiers(set="geocode1", from='2015-09-01', until='2015-09-05')
dr_list_identifiers(prefix="iso19139", from='2015-09-01', until='2015-09-20')
dr_list_identifiers(prefix="dif", from='2015-09-01', until='2015-09-20')

identifiers <- dr_list_identifiers('r')

# Data packages
identifiers[[1]]

# Data files
identifiers[[2]]
}
}

