% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_listmetadataformats.r
\name{dr_list_metadata_formats}
\alias{dr_list_metadata_formats}
\title{Get available Dryad metadata formats}
\usage{
dr_list_metadata_formats(...)
}
\arguments{
\item{...}{Curl debugging options passed on to \code{httr::GET}}
}
\value{
List of information on metadata formats.
}
\description{
Get available Dryad metadata formats
}
\examples{
\dontrun{
dr_list_metadata_formats()
}
}
