% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dryad_datasets}
\alias{dryad_datasets}
\title{List datasets}
\usage{
dryad_datasets(...)
}
\arguments{
\item{...}{Further args passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a tibble
}
\description{
List datasets
}
\examples{
\dontrun{
(x <- dryad_datasets())
x$meta
x$links
x$data
}
}
\seealso{
Other dryad-datasets: 
\code{\link{dryad_dataset_versions}()},
\code{\link{dryad_dataset}()}
}
\concept{dryad-datasets}
