% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dryad_download.R
\name{dryad_download}
\alias{dryad_download}
\title{dryad_download}
\usage{
dryad_download(dois, ...)
}
\arguments{
\item{dois}{(character) one or more DOIs, required}

\item{...}{Further args passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
file path for the file
}
\description{
Download datasets by their DOI(s)
}
\examples{
\dontrun{
dryad_download(dois = "10.5061/dryad.f385721n")
dois <- c("10.5061/dryad.f385721n", "10.5061/dryad.7ct1n", "10.5061/dryad.1g626")
dryad_download(dois = dois)
}
}
