% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{dryad_files}
\alias{dryad_files}
\title{Get metadata information about a file}
\usage{
dryad_files(ids, ...)
}
\arguments{
\item{ids}{(numeric) one or more file ids, required}

\item{...}{Further args passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of lists, each named by the input DOI
}
\description{
Get metadata information about a file
}
\examples{
\dontrun{
dryad_files(ids = 61859)
dryad_files(ids = 61858)
dryad_files(ids = c(61858, 61859))
}
}
\seealso{
Other dryad-files: 
\code{\link{dryad_files_download}()}
}
\concept{dryad-files}
