% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{dryad_files_download}
\alias{dryad_files_download}
\title{Download a specific file}
\usage{
dryad_files_download(ids, ...)
}
\arguments{
\item{ids}{(numeric) one or more file ids, required}

\item{...}{Further args passed on to \link[crul:verb-GET]{crul::verb-GET}}
}
\value{
a list of lists, each named by the input DOI
}
\description{
Download a specific file
}
\note{
there's no use of cached files here because we can't
determine what the file path will be before making the request,
which requires knowing what the data type the file will be
}
\examples{
\dontrun{
dryad_files_download(ids = 61858)
dryad_files_download(ids = 61859)
}
}
\seealso{
Other dryad-files: 
\code{\link{dryad_files}()}
}
\concept{dryad-files}
