% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\name{getalldryad_metadata}
\alias{getalldryad_metadata}
\title{Download metadata for all Dryad oai's for defined time period}
\usage{
getalldryad_metadata(...)
}
\description{
This function is defunct
}
\keyword{internal}
